Add user-agent to outgoing requests
